Add common network class for action plugin and related code refactor
