+++ lib/ansible/plugins/action/ce.py
@@ -12,7 +12,7 @@ import copy
 from ansible import constants as C
 from ansible.module_utils._text import to_text
 from ansible.module_utils.connection import Connection
-from ansible.plugins.action.normal import ActionModule as _ActionModule
+from ansible.plugins.action.network import ActionModule as ActionNetworkModule
 from ansible.module_utils.network.cloudengine.ce import ce_provider_spec
 from ansible.module_utils.network.common.utils import load_provider
 from ansible.utils.display import Display
@@ -22,11 +22,12 @@ display = Display()
 CLI_SUPPORTED_MODULES = ['ce_config', 'ce_command']
 
 
-class ActionModule(_ActionModule):
+class ActionModule(ActionNetworkModule):
 
     def run(self, tmp=None, task_vars=None):
         del tmp  # tmp no longer has any effect
 
+        self._config_module = True if self._task.action == 'ce_config' else False
         socket_path = None
 
         if self._play_context.connection == 'local':
