+++ lib/ansible/plugins/action/fortios_config.py
@@ -19,25 +19,14 @@
 from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
-import re
 
-from ansible.plugins.action.normal import ActionModule as _ActionModule
-from ansible.module_utils.network.common.backup import handle_template
-
-PRIVATE_KEYS_RE = re.compile('__.+__')
+from ansible.plugins.action.network import ActionModule as ActionNetworkModule
 
 
-class ActionModule(_ActionModule):
+class ActionModule(ActionNetworkModule):
 
     def run(self, tmp=None, task_vars=None):
-
-        if self._task.args.get('src'):
-            try:
-                self._task.args['src'] = handle_template(self)
-            except ValueError as exc:
-                return dict(failed=True, msg=exc.message)
-
-        result = super(ActionModule, self).run(tmp, task_vars)
         del tmp  # tmp no longer has any effect
 
-        return result
+        self._config_module = True
+        return super(ActionModule, self).run(task_vars=task_vars)
